Create new Origin Key Controller in API
