+++ lib/archethic_web/router.ex
@@ -71,7 +71,7 @@ defmodule ArchEthicWeb.Router do
       :last_transaction_content
     )
 
-    post("/origin_key", ArchEthicWeb.API.TransactionController, :origin_key)
+    post("/origin_key", ArchEthicWeb.API.OriginKeyController, :origin_key)
 
     post("/transaction", ArchEthicWeb.API.TransactionController, :new)
     post("/transaction_fee", ArchEthicWeb.API.TransactionController, :transaction_fee)
