+++ test/archethic_web/controllers/api/transaction_controller_test.exs
@@ -7,10 +7,6 @@ defmodule ArchEthicWeb.API.TransactionControllerTest do
   alias ArchEthic.P2P.Node
   alias ArchEthic.Crypto
 
-  alias ArchEthic.TransactionChain.Transaction
-  alias ArchEthic.TransactionChain.TransactionData
-  alias ArchEthic.TransactionChain.TransactionData.Ownership
-
   import Mox
 
   setup do
@@ -97,92 +93,6 @@ defmodule ArchEthicWeb.API.TransactionControllerTest do
                },
                "status" => "invalid"
              } = json_response(conn, 400)
-    end
-  end
-
-  describe "origin_key/2" do
-    test "should send not_found response for invalid params", %{conn: conn} do
-      conn =
-        post(conn, "/api/origin_key", %{
-          origin_public_key: "0001540315"
-        })
-
-      assert %{
-               "error" => "Invalid public key"
-             } = json_response(conn, 400)
-    end
-
-    test "should send not_found response when public key isn't found in owner transactions", %{
-      conn: conn
-    } do
-      MockDB
-      |> stub(:get_transaction, fn _, _ ->
-        {:ok,
-         %Transaction{
-           data: %TransactionData{
-             ownerships: [
-               %Ownership{
-                 secret: Base.decode16!("0001AAAAAA"),
-                 authorized_keys: %{Base.decode16!("0001BBBBBB") => Base.decode16!("0001CCCCCC")}
-               }
-             ]
-           }
-         }}
-      end)
-
-      MockDB
-      |> stub(:get_last_chain_address, fn _ ->
-        Base.decode16!("0001DDDDDD")
-      end)
-
-      conn =
-        post(conn, "/api/origin_key", %{
-          origin_public_key:
-            "00015403152aeb59b1b584d77c8f326031815674afeade8cba25f18f02737d599c39"
-        })
-
-      assert %{
-               "error" => "Public key not found"
-             } = json_response(conn, 404)
-    end
-
-    test "should send json secret values response when public key is found in owner transactions",
-         %{conn: conn} do
-      MockDB
-      |> stub(:get_transaction, fn _, _ ->
-        {:ok,
-         %Transaction{
-           data: %TransactionData{
-             ownerships: [
-               %Ownership{
-                 secret: Base.decode16!("0001AAAAAA"),
-                 authorized_keys: %{
-                   Base.decode16!(
-                     "00015403152aeb59b1b584d77c8f326031815674afeade8cba25f18f02737d599c39",
-                     case: :mixed
-                   ) => Base.decode16!("0001CCCCCC")
-                 }
-               }
-             ]
-           }
-         }}
-      end)
-
-      MockDB
-      |> stub(:get_last_chain_address, fn _ ->
-        Base.decode16!("0001DDDDDD")
-      end)
-
-      conn =
-        post(conn, "/api/origin_key", %{
-          origin_public_key:
-            "00015403152aeb59b1b584d77c8f326031815674afeade8cba25f18f02737d599c39"
-        })
-
-      assert %{
-               "encrypted_origin_private_keys" => "0001AAAAAA",
-               "encrypted_secret_key" => "0001CCCCCC"
-             } = json_response(conn, 200)
     end
   end
 end
